feat(kernel): le noyau de lot écrit à la main, et le regroupement conservé
T84. La question n'était pas « 8 ou 32 » mais : le regroupement des 21
lancers gagne-t-il encore sa complexité si le noyau est écrit à la main ?
Tout balayage antérieur mesurait autre chose — gcc ne vectorise la boucle
chaude qu'à partir de 24 (lu dans -fopt-info-vec : à largeur 8, « facteur
de déroulage 1 », donc UN accumulateur et une chaîne d'additions
dépendantes), et `bench_batch.c` prend sa largeur en variable
d'exécution.

Le noyau (`src/gn_kernel_f32.h`) tient toujours HUIT accumulateurs
vectoriels, arrangés en lignes × vecteurs selon la largeur compilée et la
cible — 8x1, 4x2, 2x4, 1x8. Accumuler R lignes contre la même colonne
casse la chaîne de dépendance et lit la colonne une fois pour R lignes ;
gcc ne tuile jamais sur les lignes de sortie. AVX/SSE2/SIMD128/scalaire.

BIT À BIT par construction et VÉRIFIÉ avant tout chronométrage :
max|Δ| = 0 aux trois largeurs, sur les deux cibles, sur les trois
noyaux. Le seul moyen de le perdre est le FMA, d'où mul et add écrites
séparément ET `-ffp-contract=off` sur l'unité de compilation.

Mesuré (décision 2-ply (0,1,3) k=12, round-robin, 3 passes) :

  natif      largeur    8      16      32
    auto (livré)      0,706   0,515   0,438 s
    intrinsèques      0,250   0,226   0,242 s   ← 11 % entre les trois
  Chromium
    auto              0,602   0,545   1,402 s
    intrinsèques      0,345   0,335   0,378 s
  Firefox
    auto              1,132   1,237   1,381 s
    intrinsèques      0,681   0,658   0,690 s

VERDICT — regroupement CONSERVÉ, largeur 32 CONSERVÉE, question CLOSE.
À noyau écrit à la main la largeur ne décide plus rien (11 % entre la
meilleure et la pire) ; le noyau décide de tout (×1,81 natif, ×3,70
Chromium, ×2,00 Firefox, à bit égal). Passer à 8 coûte 3,3 % en natif et
retirerait au passage les 93,5 % de remplissage que `rank_plays` achète.

Deux constats qui n'étaient pas prévisibles :
— l'auto-vectorisation s'effondre à largeur 32 dans V8/Chromium (facteur
  3 contre largeur 16) et PAS dans Firefox, sur le MÊME .wasm ;
— `-fassociative-math`, qui achète ×3,9 sur la passe avant scalaire (T21),
  COÛTE un facteur 2,8 sur le noyau par lot — et c'est le second chemin
  qu'une décision emprunte depuis T35. La configuration WebAssembly
  livrée est la plus lente des six mesurées.

Le changement de défaut n'est PAS livré : en natif il demande
`-march=native` (le binaire livré cible le x86-64 de base) ; en
WebAssembly le gain est disponible sans condition mais `wasm/` est en
cours de modification par T86. Interrupteur `make build
KERNEL_INTRINSICS=1` en attendant.

Instruments : bench/bench_kernel.c, bench/width_sweep.py,
bench/browser_kernel.mjs, cibles bench-width, bench-width-wasm,
bench-width-fill. Garde de non-régression : tests/test_kernel.py.

+++ bench/bench_sparsity.c
@@ -74,11 +74,11 @@ static int compare_doubles(const void *a, const void *b)
     return (x > y) - (x < y);
 }
 
-static unsigned long g_state = 20260902UL;
+static unsigned long long g_state = 20260902ULL;
 
 static int roll(void)
 {
-    g_state = g_state * 6364136223846793005UL + 1442695040888963407UL;
+    g_state = g_state * 6364136223846793005ULL + 1442695040888963407ULL;
     return (int)((g_state >> 33) % 6) + 1;
 }
 
@@ -278,7 +278,7 @@ static double paired_ratio(GnNetwork *net, GnNetwork *prune, int decisions,
         return -1.0;
     }
 
-    g_state = 20260826UL;
+    g_state = 20260826ULL;
     GnPosition pos;
     gn_position_initial(&pos);
     int done = 0;
